Extract model update metadata and make available in aggregator
